Upgrade AdminResetPassword resource to API Platform 3
