+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/AdminResetPassword.xml
@@ -1,0 +1,82 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<resources xmlns="https://api-platform.com/schema/metadata/resources-3.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="https://api-platform.com/schema/metadata/resources-3.0 https://api-platform.com/schema/metadata/resources-3.0.xsd"
+>
+    <resource
+        class="Sylius\Bundle\CoreBundle\Message\Admin\Account\ResetPassword"
+        shortName="AdminResetPassword"
+        messenger="input"
+        output="false"
+        status="202"
+    >
+        <validationContext>
+            <values>
+                <value name="groups">
+                    <values>
+                        <value>sylius</value>
+                    </values>
+                </value>
+            </values>
+        </validationContext>
+
+        <operations>
+            <operation
+                name="admin_create_reset_password"
+                class="ApiPlatform\Metadata\Post"
+                uriTemplate="/admin/reset-password-requests"
+                input="Sylius\Bundle\CoreBundle\Message\Admin\Account\RequestResetPasswordEmail"
+            >
+                <denormalizationContext>
+                    <values>
+                        <value name="groups">
+                            <values>
+                                <value>admin:reset_password:create</value>
+                            </values>
+                        </value>
+                    </values>
+                </denormalizationContext>
+                <openapiContext>
+                    <values>
+                        <value name="summary">Requests administrator's password reset</value>
+                    </values>
+                </openapiContext>
+            </operation>
+
+            <operation
+                class="ApiPlatform\Metadata\Patch"
+                name="admin_update_reset_password"
+                uriTemplate="/admin/reset-password-requests/{resetPasswordToken}"
+                input="Sylius\Bundle\CoreBundle\Message\Admin\Account\ResetPassword"
+                provider="Sylius\Bundle\ApiBundle\StateProvider\Admin\ResetPasswordProvider"
+            >
+                <denormalizationContext>
+                    <values>
+                        <value name="groups">
+                            <values>
+                                <value>admin:reset_password:update</value>
+                            </values>
+                        </value>
+                    </values>
+                </denormalizationContext>
+                <openapiContext>
+                    <values>
+                        <value name="summary">Resets administrator's password</value>
+                    </values>
+                </openapiContext>
+            </operation>
+        </operations>
+    </resource>
+</resources>
