+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/validator.xml
@@ -142,11 +142,11 @@
 <!--            <tag name="validator.constraint_validator" alias="sylius_api_unique_reviewer_email_validator" />-->
 <!--        </service>-->
 
-<!--        <service id="Sylius\Bundle\ApiBundle\Validator\Constraints\AdminResetPasswordTokenNonExpiredValidator">-->
-<!--            <argument type="service" id="sylius.repository.admin_user" />-->
-<!--            <argument type="string">%sylius.admin_user.token.password_reset.ttl%</argument>-->
-<!--            <tag name="validator.constraint_validator" alias="sylius_api_validator_admin_non_expired_password_reset_token" />-->
-<!--        </service>-->
+        <service id="Sylius\Bundle\ApiBundle\Validator\Constraints\AdminResetPasswordTokenNonExpiredValidator">
+            <argument type="service" id="sylius.repository.admin_user" />
+            <argument type="string">%sylius.admin_user.token.password_reset.ttl%</argument>
+            <tag name="validator.constraint_validator" alias="sylius_api_validator_admin_non_expired_password_reset_token" />
+        </service>
 
         <service id="Sylius\Bundle\ApiBundle\Validator\CatalogPromotion\PercentageDiscountActionValidator">
             <argument type="service" id="Sylius\Bundle\CoreBundle\SectionResolver\SectionProviderInterface" />
