+++ tests/Api/Admin/AdminResetPasswordTest.php
@@ -1,0 +1,39 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Tests\Api\Admin;
+
+use Sylius\Tests\Api\JsonApiTestCase;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+
+final class AdminResetPasswordTest extends JsonApiTestCase
+{
+    /** @test */
+    public function it_sends_administrator_password_reset_email(): void
+    {
+        $this->loadFixturesFromFile('authentication/api_administrator.yaml');
+
+        $this->client->request(
+            method: Request::METHOD_POST,
+            uri: '/api/v2/admin/reset-password-requests',
+            server: self::CONTENT_TYPE_HEADER,
+            content: json_encode([
+                'email' => 'api@example.com',
+            ], JSON_THROW_ON_ERROR),
+        );
+
+        $response = $this->client->getResponse();
+        $this->assertResponseCode($response, Response::HTTP_ACCEPTED);
+    }
+}
